requirements.txt
# AI Hedge Fund Dependencies

# OpenAI for AI agents
openai==1.12.0

# Data handling
pandas==2.1.1
numpy==1.26.0

# Market data
yfinance==0.2.31
alpha_vantage==2.3.1

# Paper trading
alpaca-trade-api==3.0.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Visualization and dashboard
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.18.0
dash==2.14.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1 